get-dataset/requirements.txt
# Function dependencies, for example:
# package>=version
wget>=3.2
google-cloud-storage>=1.42.3